+++ packages/gateway/src/api/config/config.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+
+import { ConfigController } from './config.controller';
+
+@Module({
+  controllers: [ConfigController],
+})
+export class ConfigAPIModule {}