+++ packages/gateway/src/auth/cookie.strategy.ts
@@ -1,8 +1,11 @@
+import _ from 'lodash';
 import { Strategy } from 'passport-cookie';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException, Req } from '@nestjs/common';
 import { Request } from 'express';
 
+import { Role } from '../constants/role';
+import { UserContext } from '../types/app';
 import { AuthService } from './auth.service';
 
 @Injectable()
@@ -18,11 +21,23 @@ export class CookieStrategy extends PassportStrategy(Strategy) {
   public async validate(
     @Req() req: Request,
     accessToken: string
-  ): Promise<{ readonly accessToken: string }> {
-    const result = await this.authService.validateAccessToken(accessToken);
-    if (!result) {
+  ): Promise<UserContext> {
+    const isAccessToken = this.authService.validateAccessToken(accessToken);
+    const isViewerToken = this.authService.validateViewerToken(accessToken);
+    const roles: UserContext['roles'] = [];
+
+    if (!isAccessToken && !isViewerToken) {
       throw new UnauthorizedException();
     }
-    return { accessToken };
+
+    if (isViewerToken) {
+      roles.push(Role.VIEWER);
+    }
+
+    if (isAccessToken) {
+      roles.push(Role.ADMIN, Role.VIEWER);
+    }
+
+    return { accessToken, roles: _.uniq(roles) };
   }
 }