+++ packages/gateway/src/surgio/surgio.service.ts
@@ -3,11 +3,11 @@ import { Artifact } from 'surgio/build/generator/artifact';
 import { PossibleProviderType } from 'surgio/build/provider';
 import { CommandConfig } from 'surgio/build/types';
 
-import { SurgioHelper } from './surgio-helper';
+import { KEY, SurgioHelper } from './surgio-helper';
 
 @Injectable()
 export class SurgioService {
-  constructor(@Inject('SURGIO_HELPER') public surgioHelper: SurgioHelper) {}
+  constructor(@Inject(KEY) public surgioHelper: SurgioHelper) {}
 
   public get config(): CommandConfig {
     return this.surgioHelper.config;